Move fs commands to fs module and migrate iOS cmds to darwin module
